Move get_sitk_image method back
